+++ src/muon/atac/tl.py
@@ -685,7 +685,7 @@ def count_fragments_features(
     stranded: bool = False,
     extend_upstream: float = 2e3,
     extend_downstream: int = 0,
-    count_reads: bool = True,
+    count_reads: bool = False,
 ) -> AnnData:
     """Count fragments overlapping given Features. Returns cells x features matrix.
 
@@ -700,11 +700,8 @@ def count_fragments_features(
             according to each feature's strand information.
         extend_upstream: Number of nucleotides to extend every gene upstream (2000 by default to extend gene coordinates to promoter regions)
         extend_downstream: Number of nucleotides to extend every gene downstream (0 by default)
-        count_reads: NOTE: default will be changed to False from v0.2.
-            If to count reads instead of fragments.
-            If True, the number of reads (read support) per fragment will be used.
-            This will also include duplicate read pairs.
-            If False, `1` will be added for each fragment.
+        count_reads: Whether to count reads instead of fragments. If `True`, the number of reads (read support) per fragment will be used.
+            This will also include duplicate read pairs. If `False`, `1` will be added for each fragment.
     """
     if isinstance(data, AnnData):
         adata = data
@@ -731,13 +728,6 @@ def count_fragments_features(
         raise ImportError(
             "pysam is not available. It is required to work with the fragments file. Install pysam from PyPI (`pip install pysam`) or from GitHub (`pip install git+https://github.com/pysam-developers/pysam`)"
         ) from None
-
-    if count_reads:
-        warn(
-            "From v0.2, by default, unique fragments will be counted instead of reads. See muon#110 for details.",
-            FutureWarning,
-            stacklevel=2,
-        )
 
     n = adata.n_obs
     n_features = features.shape[0]